opensearch build & configure

+++ lxc/scripts/opensearch/build_data_node.sh
@@ -3,50 +3,49 @@
 # "Install OpenSeach from an APT repository"
 
 # Mounts an external data volume into container (for a data node)
-
-#############################################################################
-#
-# PRIOR TO RUNNING THIS SCRIPT 
-#
-# Calculate password hashes for admin and invenio_usr & copy these into internal_users.yml
-#
-#############################################################################
 NAME=$1 #"opensearch_d1"
-CONFIG_ROOT=$2 #"../lxc_config.conf"
-DATA_MNT=$3 #"/home/leeb/.local/var/lxc/opensearch_1/data"
-LOG_MNT=$4  #"/home/leeb/.local/var/lxc/opensearch_1/log"
+DATA_MNT=$2 #"/home/leeb/.local/var/lxc/opensearch_1/data"
+LOG_MNT=$3 #"/home/leeb/.local/var/lxc/opensearch_1/log"
 
-export OPENSEARCH_VERSION='2.15.0'
-export GPG_SIGNATURE='c5b7 4989 65ef d1c2 924b a9d5 39d3 1987 9310 d3fc'
-export OPENSEARCH_INITIAL_ADMIN_PASSWORD='its_@_Secret123'
+OPENSEARCH_VERSION='2.15.0'
+GPG_SIGNATURE='c5b7 4989 65ef d1c2 924b a9d5 39d3 1987 9310 d3fc'
 
-echo $'\n'"lxc.mount.entry = ${PWD}/mount home/host none bind,create=dir 0 0"\
+MOUNT="${PREFIX}/services/opensearch/data-node"
+
+if [ ! -f ${NAME}.conf ]
+then
+echo $'\n'"lxc.mount.entry = ${MOUNT} home/host none bind,create=dir 0 0"\
      $'\n'"lxc.mount.entry = ${DATA_MNT} var/opensearch/data none bind,create=dir 0 0"\
      $'\n'"lxc.mount.entry = ${LOG_MNT} var/log/opensearch none bind,create=dir 0 0"\
-    | cat ${CONFIG_ROOT} -\
+    | cat ${LXC_CONFIG} -\
     | tee -a ${NAME}.conf
+else
+    echo "Config file already exists, moving on"
+fi
 
-source ../create_container.sh
-source ../secrets.sh
-
-bash create_self_signed_ssl_certs.sh mount/ssl
 
 create_container ${NAME} ${NAME}.conf && \
-    rm -f ${NAME}.conf && \ 
+    
+    lxc_start -n ${NAME} && \
+
+    cp -R ssl/ ${MOUNT} && \
+    
+    lxc_attach -n ${NAME} --clear-env -- \
+	       /home/host/scripts/build.sh ${OPENSEARCH_INITIAL_ADMIN_PASSWORD} \
+	       ${OPENSEARCH_VERSION} ${GPG_SIGNATURE} && \
+
+    lxc_attach -n ${NAME} --clear-env -- \
+	       /home/host/scripts/configure.sh ${OPENSEARCH_ADMIN_PASSWD} \
+	       ${OPENSEARCH_AEDATASTORE_PASSWD} && \
+
+    sed -ir '/^lxc.mount.entry.*home\/host/d' ${LXC_UNPRIV_DIR}/${NAME}/config && \
+
+    lxc-stop -n ${NAME} && lxc_start -n ${NAME}
 
 
-systemd-run --user --scope -p "Delegate=yes" -- lxc-start -n ${NAME} && \
 
-systemd-run --user --scope -p "Delegate=yes" -- lxc-attach -n ${NAME} --clear-env \
-	    --keep-var OPENSEARCH_INITIAL_ADMIN_PASSWORD \
-	    --keep-var OPENSEARCH_VERSION \
-	    --keep-var GPG_SIGNATURE \
-	    -- /home/host/scripts/data-node/build.sh
 
-systemd-run --user --scope -p "Delegate=yes" -- lxc-attach -n ${NAME} --clear-env \
-	    --keep-var OPENSEARCH_ADMIN_PASSWD \
-	    --keep-var OPENSEARCH_AEDATASTORE_PASSWD \
-	    -- /home/host/scripts/data-node/configure.sh
+
